minor fix in embedder input

+++ src/embedders.py
@@ -179,7 +179,7 @@ class TransformerEmbeddings(EmbedderBase):
     def __enter__(self):
         if not self._model:         
             from sentence_transformers import SentenceTransformer
-            self._model = SentenceTransformer(self.model_path, processor_kwargs=self.tokenizer_kwargs, device=self.device, dtype=torch.bfloat16)
+            self._model = SentenceTransformer(self.model_path, processor_kwargs=self.tokenizer_kwargs, device=self.device)
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
